+++ sql/utils/sql_utils.py
@@ -414,3 +414,5 @@ def filter_db_list(db_list, db_name_regex: str, is_match_regex: bool, key="value
         if (is_match_regex and is_match) or (not is_match_regex and not is_match):
             filtered_list.append(db)
     return filtered_list
+
+
